# OpenAPI.yml
openapi: '3.0.3'
info:
  title: 'Notifications Server'
  version: '1.0'
servers:
  - url: 185.92.192.88:3000/api
    description: Production server
  - url: 185.92.192.88:3001/dev/api
    description: Development server
paths:
  /devices:
    post:
      summary: Add a new device.
      description: Inserts a new device record into the database.
      tags:
        - Device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceRequestBody'           
      responses:
        '200':
          description: Device added successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  id:
                    type: integer
                    example: 1
        '400':
          description: Duplicate device data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Duplicate found: phone number or token already exists."
        '500':
          description: Unknown error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnknownError'

    get:
      summary: Get all devices.
      description: Retrieves a list of all devices.
      tags:
        - Device
      responses:
        '200':
          description: List of devices retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  devices:
                    $ref: '#/components/schemas/Devices'
        '500':
          description: Error while retrieving devices.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnknownError'

  /devices/{id}:
    get:
      summary: Get device by ID.
      description: Retrieves details of a specific device by ID.
      tags:
        - Device
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the device to retrieve.
      responses:
        '200':
          description: Device retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  device:
                    oneOf:
                      - $ref: '#/components/schemas/Device'
                      - type: string
                        example: No device found.
        '500':
          description: Error while retrieving the device.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnknownError'

    put:
      summary: Update a device.
      description: Updates a device's details by ID.
      tags:
        - Device
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the device to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceRequestBody'
      responses:
        '200':
          description: Device updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  id:
                    type: integer
                    example: 1
        '404':
          description: Device not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Device not found.
        '400':
          description: Duplicate device data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Duplicate found: phone number or token already exists."
        '500':
          description: Unknown error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnknownError'

    delete:
      summary: Delete a device.
      description: Deletes a device by its ID.
      tags:
        - Device
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the device to delete.
      responses:
        '200':
          description: Device deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Device deleted successfully.
        '404':
          description: Device not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Device not found.
        '400':
          description: Invalid ID format.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Wrong input id format.
        '500':
          description: Unknown error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnknownError'

  /notifications/send:
    post:
      summary: Send a notification.
      description: Sends a notification to a mobile device.
      tags:
        - Notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - identifier
              properties:
                identifier:
                  type: string
                  description: |
                    A unique identifier consisting of exactly 8 hexadecimal characters.
                    Valid characters include numbers (0-9) and letters (A-F, a-f).
                  example: "3A0DE20B"
                  pattern: "^[A-Fa-f0-9]{8}$"
                cid:
                  type: string
                  description: |
                    CID message format consisting of exactly 20 numeric characters.
                  example: "00050001101412004198"
                  pattern: "^[0-9]{20}$"

      responses:
        '200':
          description: Notification send success,
          content: 
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Successfully sent message.
        '400':
          description: Invalid FCM registration token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Invalid FCM registration token provided. Please check the token.
        '500':
          description: Unexpected error while sending the notification.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnknownError'

components:
  schemas:
    DeviceRequestBody:
      type: object
      required:
      - identifier
      - token
      properties:
        token:
          type: string
          example: "abcd1234efgh5678"
        identifier:
          type: string
          description: |
            A unique identifier consisting of exactly 8 hexadecimal characters.
            Valid characters include numbers (0-9) and letters (A-F, a-f).
          example: "3A0DE20B"
          pattern: "^[A-Fa-f0-9]{8}$"
        
    Device:
      type: object
      required:
      - id
      - identifier
      - token
      properties:
        id:
          type: integer
          example: 1
        token:
          type: string
          example: "abcd1234efgh5678"
        identifier:
          type: string
          description: |
            A unique identifier consisting of exactly 8 hexadecimal characters.
            Valid characters include numbers (0-9) and letters (A-F, a-f).
          example: "3A0DE20B"
          pattern: "^[A-Fa-f0-9]{8}$"
    Devices:
      type: array
      items:
        $ref: '#/components/schemas/Device'
    UnknownError:
      type: object
      required:
        - success
        - message
        - error
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: An unexpected error occurred while sending the notification.
        error:
          type: string
          example: "Error details."